+++ dms-enterprise-20181101/alibabacloud_dms_enterprise20181101/models/_create_order_shrink_request.py
@@ -11,6 +11,7 @@ class CreateOrderShrinkRequest(DaraModel):
         comment: str = None,
         plugin_param_shrink: str = None,
         plugin_type: str = None,
+        real_login_user_uid: str = None,
         related_user_list: str = None,
         tid: int = None,
     ):
@@ -28,6 +29,7 @@ class CreateOrderShrinkRequest(DaraModel):
         # 
         # This parameter is required.
         self.plugin_type = plugin_type
+        self.real_login_user_uid = real_login_user_uid
         # The IDs of the stakeholders that are involved in the ticket. Separate multiple IDs with commas (,).
         self.related_user_list = related_user_list
         # The ID of the tenant. You can call the [GetUserActiveTenant](https://help.aliyun.com/document_detail/198073.html) or [ListUserTenants](https://help.aliyun.com/document_detail/198074.html) operation to obtain the tenant ID.
@@ -53,6 +55,9 @@ class CreateOrderShrinkRequest(DaraModel):
         if self.plugin_type is not None:
             result['PluginType'] = self.plugin_type
 
+        if self.real_login_user_uid is not None:
+            result['RealLoginUserUid'] = self.real_login_user_uid
+
         if self.related_user_list is not None:
             result['RelatedUserList'] = self.related_user_list
 
@@ -74,6 +79,9 @@ class CreateOrderShrinkRequest(DaraModel):
 
         if m.get('PluginType') is not None:
             self.plugin_type = m.get('PluginType')
+
+        if m.get('RealLoginUserUid') is not None:
+            self.real_login_user_uid = m.get('RealLoginUserUid')
 
         if m.get('RelatedUserList') is not None:
             self.related_user_list = m.get('RelatedUserList')